requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
distro==1.9.0
exceptiongroup==1.2.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
kombu==5.3.4
mysql-connector==2.2.9
mysql-connector-python==8.2.0
openai==1.7.2
prompt-toolkit==3.0.43
protobuf==4.21.12
pyaes==1.6.1
pyasn1==0.5.1
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
python-dotenv==1.0.0
regex==2023.12.25
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
Telethon==1.33.1
tiktoken==0.5.2
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
